% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_prep_soils_ssurgo.R
\name{dssat_prep_soils_ssurgo}
\alias{dssat_prep_soils_ssurgo}
\title{Prepare soils imported from the NRCS SSURGO database for use in DSSAT}
\usage{
dssat_prep_soils_ssurgo(ssurgo)
}
\arguments{
\item{ssurgo}{A NRCS SSURGO dataset, created using \code{\link{FedData::get_ssurgo()}}.}
}
\value{
A named list of SSURGO data:
\enumerate{
\item 'mapunits': A \code{SpatialPolygonsDataFrame} of mapunits of each soil,
\item 'components': A \code{\link{tibble}} of soil component-level data,
\item 'horizons': A \code{\link{tibble}} of soil horizon-level data.
}
}
\description{
This function calculates the soil characteristics needed to run a DSSAT model
from an NRCS SSURGO dataset, created using \code{\link{FedData::get_ssurgo()}}.
}
