% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_run_batch.R
\name{dssat_run_batch}
\alias{dssat_run_batch}
\title{Run a batch spatial dssat analysis from a
cultigen, weather, and soil object.}
\usage{
dssat_run_batch(name,
  cultivars = dssat_read_cultigen("~/DSSAT46/Genotype/MZCER046.CUL") \%>\%
  dplyr::filter(`@VAR#` \%in\% c("AC0001", "GF0001", "GF0101", "GF0201",
  "GF0301", "GF0401", "GF0501")), weather, soil, earliest_planting_doy = 136,
  latest_planting_doy = 167, output_dir = tempfile())
}
\arguments{
\item{name}{}

\item{cultivars}{}

\item{weather}{}

\item{soil}{}

\item{earliest_planting_doy}{}

\item{latest_planting_doy}{}

\item{output_dir}{}
}
\value{
A list containing field, cultivar, and yield information.
}
\description{
Run a batch spatial dssat analysis from a
cultigen, weather, and soil object.
}
