% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssat_spatial.R
\name{dssat_spatial}
\alias{dssat_spatial}
\title{Run a spatial DSSAT model from FedData downloads}
\usage{
dssat_spatial(template, label, output.dir = "./OUTPUT/",
  raw.dir = paste0(output.dir, "/DATA/RAW/"),
  extraction.dir = paste0(output.dir, "/DATA/EXTRACTIONS/"),
  force.redo = FALSE)
}
\arguments{
\item{template}{A Raster* or Spatial* object to serve
as a template for cropping.}

\item{label}{A character string naming the study area.}

\item{output.dir}{A character string indicating a directory where data should be output.}

\item{raw.dir}{A character string indicating where raw downloaded files should be put.
The directory will be created if missing. Defaults to \code{[output.dir]/DATA/RAW/}.}

\item{extraction.dir}{A character string indicating where the extracted and cropped SSURGO shapefiles should be put.
The directory will be created if missing. Defaults to \code{[output.dir]/DATA/EXTRACTIONS/}.}

\item{force.redo}{If an extraction for this template and label already exists, should a new one be created? Defaults to FALSE.}
}
\value{
A named list containing "spatial" and "tabular" data.
}
\description{
This function reads in several gridded datasets necessary for running a
spatial DSSAT model, selects the resolution to run the model at based on
the lowest resolution dataset (Daymet), aggregates all other datasets
down to that resolution, and preps and runs the simulation on a per-cell basis
}

