% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daymet.R
\name{get_daymet}
\alias{get_daymet}
\title{Download and crop the 1-km DAYMET daily weather dataset.}
\usage{
get_daymet(x, years = 1980:2016, vars = c("tmax", "tmin", "srad", "vp",
  "swe", "prcp", "dayl"))
}
\arguments{
\item{x}{An object class \code{sf} to serve as a template}

\item{years}{A numeric vector of years to extract.}

\item{vars}{A character vector of variables to extract.\cr
The available variables are:
\describe{
\item{dayl}{Duration of the daylight period in seconds per day.
This calculation is based on the period of the day during
which the sun is above a hypothetical flat horizon.}
\item{prcp}{Daily total precipitation in millimeters per day,
sum of all forms converted to water-equivalent.}
\item{srad}{Incident shortwave radiation flux density in watts per square meter,
taken as an average over the daylight period of the day.\cr
NOTE: Daily total radiation (MJ/m2/day) can be calculated as follows:\cr
\deqn{\frac{srad \times dayl}{l000000}}{srad*dayl/l000000}}
\item{swe}{Snow water equivalent in kilograms per square meter.
The amount of water contained within the snowpack.}
\item{tmax}{Daily maximum 2-meter air temperature in degrees Celsius.}
\item{tmin}{Daily minimum 2-meter air temperature in degrees Celsius.}
\item{vp}{Water vapor pressure in pascals. Daily average partial pressure of water vapor.}
}}
}
\description{
This function purposefully masks the \code{\link[FedData:get_daymet]{FedData::get_daymet()}} function,
and will likely replace it eventually.
}
