% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sf.R
\name{st_as_sf}
\alias{st_as_sf}
\alias{st_as_sf.Raster}
\title{Convert a Raster* object to an sf object}
\usage{
\method{st_as_sf}{Raster}(x, ...)
}
\arguments{
\item{x}{Raster* to be converted into an object class \code{sf}}

\item{...}{further specifications, see \link{st_as_sf}}
}
\description{
Convert a Raster* object to an sf object.
This does not preserve the data, but only converts
the grid to an sf object with geometry type POLYGONS.
}
